+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 
 - Support parentheses in `test` / `[` commands in container startup scripts,
   via dependency update of mvdan.cc/sh.
+- When user requests a bind of `/dev:/dev` or `/dev/xxx:/dev/xxx` in OCI-mode,
+  ensure that it is bind mounted with appropriate flags so that it is usable in
+  the container.
 
 ## 4.0.1 \[2023-10-13\]